All modules /W3 clean (vs2015)

+++ code/Modules/IMUI/IMUI.cc
@@ -31,7 +31,7 @@ IMUI::Discard() {
 void
 IMUI::NewFrame(Duration frameDuration) {
     o_assert_dbg(IsValid());
-    state->imguiWrapper.NewFrame(frameDuration.AsSeconds());
+    state->imguiWrapper.NewFrame((float32)frameDuration.AsSeconds());
 }
 
 //------------------------------------------------------------------------------
